1.Added constraint to data model 2.Added logOut button and functionality

+++ FitnessApp/Models/ProfileModel.swift
@@ -12,8 +12,8 @@ struct ProfileModel {
     var email: String
     var name: String
     var age: Int
-    var weight: Int
-    var growth: Int
+    var weight: Double
+    var growth: Double
     var gender: String
     var profileImagePath: String
 }
